pyproject.toml
[project]
name = "pydantic-tracing"
version = "0.1.0"
description = "Simple agent with pydantic-ai, Temporal workflows, and LangSmith evaluation"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=1.40.0",
    "temporalio>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "langsmith>=0.6.1",
    "langchain-openai>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["agent.py", "workflow.py", "run.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
